+++ assignments/corejava/bankingapp-with-arrays/src/com/indium/bankingapp/Service/BankOper.java
@@ -3,12 +3,9 @@ package com.indium.bankingapp.Service;
 import com.indium.bankingapp.Model.Account;
 
 public interface BankOper {
-
     void createAccount(Account account);
     Account[] getAllAccounts();
     Account getAccountById(int id);
-    void updateBalance(Account account, double newBalance);
+    void updateBalance(Account account, int newBalance);
     void deleteAccount(int id);
 }
-
-
